+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
+## [0.4.0.0] 
+- Changed the system to create orders once the customer returns from the
+Divido Payment Platform
+
 ## [0.3.1.0] 
 - Fix bug with not recognising Api key
 - Fix bug with broken html tag
-- Fix bug to chechkout with vouchers 
+- Fix bug to checkout with vouchers 
 
 ## [0.3.0.0] 
 - Functionality to offer bespoke finance plans on a per product basis